lance_utils/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-ai"
version = "0.1.0"
description = "A flexible, annotation-first library for building Lance datasets."
authors = [{ name = "Gemini" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyarrow",
    "pandas",
    "Pillow",
    "matplotlib",
    "rich",
    "click",
    "numpy",
    "tqdm",
    "pydantic",
    "pylance",
    "psutil",
]

[project.scripts]
atlas = "atlas.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["atlas*"]
